+++ migrations/20191010114626_locations_and_sources_owners_views.down.sql
@@ -1,0 +1,7 @@
+--- revoke grants
+REVOKE SELECT ON eventzimmer.sources_by_owner FROM organizer;
+REVOKE SELECT ON eventzimmer.locations_by_owner FROM organizer;
+
+--- drop views
+DROP VIEW eventzimmer.sources_by_owner;
+DROP VIEW eventzimmer.locations_by_owner;